Put SelfCallKind in a separate file
